+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- cmd/udns: add rrset query tool
 
 ## [1.2.1] - 2016-06-13
 ### Fixed